+++ websocketpp/impl/endpoint_impl.hpp
@@ -87,8 +87,7 @@ endpoint<connection,config>::create_connection() {
 }
 
 template <typename connection, typename config>
-void endpoint<connection,config>::interrupt(connection_hdl hdl,
-    lib::error_code & ec)
+void endpoint<connection,config>::interrupt(connection_hdl hdl, lib::error_code & ec)
 {
     connection_ptr con = get_con_from_hdl(hdl,ec);
     if (ec) {return;}
@@ -106,8 +105,42 @@ void endpoint<connection,config>::interrupt(connection_hdl hdl) {
 }
 
 template <typename connection, typename config>
-void endpoint<connection,config>::send(connection_hdl hdl, std::string const &
-    payload, frame::opcode::value op, lib::error_code & ec)
+void endpoint<connection,config>::pause_reading(connection_hdl hdl, lib::error_code & ec)
+{
+    connection_ptr con = get_con_from_hdl(hdl,ec);
+    if (ec) {return;}
+
+    ec = con->pause_reading();
+}
+
+template <typename connection, typename config>
+void endpoint<connection,config>::pause_reading(connection_hdl hdl) {
+    lib::error_code ec;
+    pause_reading(hdl,ec);
+    if (ec) { throw ec; }
+}
+
+template <typename connection, typename config>
+void endpoint<connection,config>::resume_reading(connection_hdl hdl, lib::error_code & ec)
+{
+    connection_ptr con = get_con_from_hdl(hdl,ec);
+    if (ec) {return;}
+
+    ec = con->resume_reading();
+}
+
+template <typename connection, typename config>
+void endpoint<connection,config>::resume_reading(connection_hdl hdl) {
+    lib::error_code ec;
+    resume_reading(hdl,ec);
+    if (ec) { throw ec; }
+}
+
+
+
+template <typename connection, typename config>
+void endpoint<connection,config>::send(connection_hdl hdl, std::string const & payload, 
+    frame::opcode::value op, lib::error_code & ec)
 {
     connection_ptr con = get_con_from_hdl(hdl,ec);
     if (ec) {return;}
@@ -116,8 +149,8 @@ void endpoint<connection,config>::send(connection_hdl hdl, std::string const &
 }
 
 template <typename connection, typename config>
-void endpoint<connection,config>::send(connection_hdl hdl, std::string const &
-    payload, frame::opcode::value op)
+void endpoint<connection,config>::send(connection_hdl hdl, std::string const & payload, 
+    frame::opcode::value op)
 {
     lib::error_code ec;
     send(hdl,payload,op,ec);
@@ -187,8 +220,7 @@ void endpoint<connection,config>::ping(connection_hdl hdl, std::string const &
 }
 
 template <typename connection, typename config>
-void endpoint<connection,config>::ping(connection_hdl hdl, std::string const &
-    payload)
+void endpoint<connection,config>::ping(connection_hdl hdl, std::string const & payload)
 {
     lib::error_code ec;
     ping(hdl,payload,ec);
@@ -196,8 +228,8 @@ void endpoint<connection,config>::ping(connection_hdl hdl, std::string const &
 }
 
 template <typename connection, typename config>
-void endpoint<connection,config>::pong(connection_hdl hdl, std::string const &
-    payload, lib::error_code & ec)
+void endpoint<connection,config>::pong(connection_hdl hdl, std::string const & payload, 
+    lib::error_code & ec)
 {
     connection_ptr con = get_con_from_hdl(hdl,ec);
     if (ec) {return;}
@@ -205,8 +237,7 @@ void endpoint<connection,config>::pong(connection_hdl hdl, std::string const &
 }
 
 template <typename connection, typename config>
-void endpoint<connection,config>::pong(connection_hdl hdl, std::string const &
-    payload)
+void endpoint<connection,config>::pong(connection_hdl hdl, std::string const & payload)
 {
     lib::error_code ec;
     pong(hdl,payload,ec);